A few bug fix
For the case the dcache is described in the xml file, we need to avoid StandardEventBuilder call data->process() because dcache is not supported in this context.
If data->process() is called, it generated an 'seg fault' because data is nullptr when dcache is described and NoCacheDataAccessBuilder if not.
To prevent this, we need to check if data is not nullptr before calling the process function.
PS : for dcache support, use otawa-dcache